(VDB-1082) Update primary keys for large tables to bigint

+++ db/migrations/00008_create_event_logs_table.sql
@@ -1,9 +1,9 @@
 -- +goose Up
 CREATE TABLE public.event_logs
 (
-    id           SERIAL PRIMARY KEY,
+    id           BIGSERIAL PRIMARY KEY,
     header_id    INTEGER NOT NULL REFERENCES public.headers (id) ON DELETE CASCADE,
-    address      INTEGER NOT NULL REFERENCES public.addresses (id) ON DELETE CASCADE,
+    address      BIGINT  NOT NULL REFERENCES public.addresses (id) ON DELETE CASCADE,
     topics       BYTEA[],
     data         BYTEA,
     block_number BIGINT,
